futbot-v3/pyproject.toml
[project]
name = "test-robot"
version = "0.1.0"
requires-python = ">=3.11,<3.15"

dependencies = [
    "opencv-python-headless>=4.13.0.92,<5",
    "numpy>=2,<3",
    "pyserial>=3.5",
    "smbus2>=0.4",
    "onnxruntime>=1.24.4,<2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.uv]
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
]
